Fiosynth - disable LM debug log
Summary: Currently test enables LM with debug log enabled with value 0x1, which means LM debug log will only trigger when the read counter for bucket 0 is incremented. All of the above test have 5ms as the threshold for bucket 0, so we are seeing the LM debug log trigger unnecessarily, which adds latency to host IO during operation.

Reviewed By: darryleg

Differential Revision: D51140163

fbshipit-source-id: e7f81ff25b84eb4243566c85a6d1fcb36cf8b4c9