Bug fix in nvme list output conversion to old format
Summary:
Previously the nvme list -o json output from newer nvme version was converted to old format but it was only considering 1 namspace per drive which caused some tests to fail.

To fix this the conversion function is updated to loop through all the namespaces and return the list of cmd output converted to old format.

Reviewed By: lclarke1

Differential Revision: D76545885

fbshipit-source-id: 295d39435cf000bcac90bea259be4f5dc90c3ac4